# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blossom Buddy API",
        "description": "API pour la gestion des plantes et utilisateurs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/plant": {
            "post": {
                "tags": [
                    "User Plants"
                ],
                "summary": "Ajouter une plante à un utilisateur",
                "operationId": "f209da9110e8cae9cfd01adbdc4fba02",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "common_name",
                                    "city"
                                ],
                                "properties": {
                                    "common_name": {
                                        "type": "string",
                                        "example": "Ficus"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Paris"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plante ajoutée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plante ajoutée à l'utilisateur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plante non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Plante non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/plants": {
            "get": {
                "tags": [
                    "User Plants"
                ],
                "summary": "Récupérer toutes les plantes de l'utilisateur",
                "operationId": "e373eeb86287c1615c861b9daff073fa",
                "responses": {
                    "200": {
                        "description": "Liste des plantes de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "common_name": {
                                                "type": "string",
                                                "example": "Ficus"
                                            },
                                            "watering_general_benchmark": {
                                                "type": "object"
                                            },
                                            "pivot": {
                                                "properties": {
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Paris"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/plant/{id}": {
            "delete": {
                "tags": [
                    "User Plants"
                ],
                "summary": "Supprimer une plante de l'utilisateur",
                "operationId": "ee358d3a567b9f3c3986f4d2686254bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relation plante-utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plante supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plante retirée de la liste de l'utilisateur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relation plante-utilisateur non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Relation plante-utilisateur non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}